Update to prereq bullet #5
I've split prerequisite bullet #5 into two: First one is about the instance type only. Here I added a sentence stating where the error will be shown. 2nd bullet is the one about the different controls and what they depend on.
My motivation for this split comes from customers struggling to notice that instance types (B2C vs B2B) is a major factor for the card add-in. 
I considered grouping the 2nd bullet with in the first, but I doubt it makes a difference worth the cost of having such deep bullet points.